requirements.txt
gradio
scikit-learn
numpy<2
pandas
scipy
scikit-surprise
kaggle
kagglehub